+++ CHANGELOG.md
@@ -7,6 +7,7 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 ## Unreleased
 ### Added
 - Unified solution for preprint words
+- Added preprint tombstone page.
 
 ### Changed
 - Styling and format of the branded navbar to match current styling in ember-osf