+++ CHANGELOG.md
@@ -21,6 +21,7 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 - Sharing from the preprint detail view to use ember-osf sharing icons
 - Update preprints to work with a new unified version of ember-osf (works with both preprints and reviews apps)
 - Update preprints to use ember-cli@2.18
+- User search to also search profile information
 
 ## [0.118.7] - 2018-06-21
 ### Added