+++ CHANGELOG.md
@@ -4,6 +4,10 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](http://keepachangelog.com/en/1.0.0/)
 and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.html).
 
+## [0.119.0] - 2018-08-16
+### Added
+- Hypothes.is commenting toggle
+
 ## [0.118.8] - 2018-08-11
 ### Changed
 - Updated to use latest provider assets